moves copy to clipboard util methods into mixin removes component specific internal variables from mixin adds data and columnNames arguments to copyTableData method adds clipboardMixin to DataEditorTab component adds clipboard functionality to tabulator adds context menus to data editor cells
