+++ pgmanage/app/static/pgmanage_frontend/src/components/QueryResultTabs.vue
@@ -79,13 +79,14 @@ import isEmpty from 'lodash/isEmpty';
 import mean from 'lodash/mean';
 import last from 'lodash/last';
 import { Tab } from "bootstrap";
-import { handleError } from "../logging/utils";
 import dialects from './dialect-data';
+import ClipboardMixin from "../mixins/table_clipboard_copy_mixin";
 
 export default {
   components: {
     ExplainTabContent,
   },
+  mixins: [ClipboardMixin,],
   props: {
     tabId: String,
     workspaceId: String,
@@ -223,115 +224,6 @@ export default {
     this.handleResize();
   },
   methods: {
-    copyTableData(format) {
-      const data = last(this.table.getRangesData());
-
-      let headers = [];
-      let headerIndexMap = {}; // Map header titles to their original indices
-      let nameCount = {}; // to track duplicates
-
-
-      Object.keys(last(data)).forEach((key) => {
-          const originalIndex = parseInt(key, 10);
-          let  header = this.columns[originalIndex];
-
-          // Track duplicates
-          if (nameCount[header]) {
-            nameCount[header]++;
-            header = `${header}_${nameCount[header]}`;
-          } else {
-            nameCount[header] = 1;
-          }
-
-          headers.push(header);
-          headerIndexMap[header] = originalIndex;
-      });
-
-      if (format === "json") {
-        const jsonOutput = this.generateJson(data, headers, headerIndexMap);
-        this.copyToClipboard(jsonOutput);
-      } else if (format === "csv") {
-        const csvOutput = this.generateCsv(data, headers, headerIndexMap);
-        this.copyToClipboard(csvOutput);
-      } else if (format === "markdown") {
-        const markdownOutput = this.generateMarkdown(data, headers, headerIndexMap);
-        this.copyToClipboard(markdownOutput);
-      }
-    },
-    copyToClipboard(text) {
-      navigator.clipboard
-        .writeText(text)
-        .then(() => {
-        })
-        .catch((error) => {
-          handleError(error);
-        });
-    },
-    generateJson(data, headers, headerIndexMap) {
-      const mappedData = data.map((row) => {
-        const mappedRow = {};
-        headers.forEach((header) => {
-          const originalIndex = headerIndexMap[header];
-          mappedRow[header] = row[originalIndex];
-        });
-        return mappedRow;
-      });
-
-      return JSON.stringify(mappedData, null, 2);
-    },
-    generateCsv(data, headers, headerIndexMap) {
-      const csvRows = [];
-
-      // Add header row
-      csvRows.push(headers.join(settingsStore.csvDelimiter));
-
-      data.forEach((row) => {
-        const rowValues = headers.map((header) => {
-          const originalIndex = headerIndexMap[header];
-          return row[originalIndex] || ""; 
-        })
-        csvRows.push(rowValues.join(settingsStore.csvDelimiter));
-      });
-
-      return csvRows.join("\n");
-    },
-    generateMarkdown(data, headers, headerIndexMap) {
-      const columnWidths = headers.map((header) => {
-        const originalIndex = headerIndexMap[header];
-        const maxDataLength = data.reduce(
-          (max, row) => Math.max(max, (row[originalIndex] || "").toString().length),
-          0
-        );
-        return Math.max(header.length, maxDataLength);
-      });
-
-      // Helper to pad strings to a given length
-      const pad = (str, length) => str.toString().padEnd(length, " ");
-
-      const mdRows = [];
-
-      // Add padded header row
-      mdRows.push(
-        `| ${headers
-          .map((header, index) => pad(header, columnWidths[index]))
-          .join(" | ")} |`
-      );
-
-      // Add separator row
-      mdRows.push(
-        `| ${columnWidths.map((width) => "-".repeat(width)).join(" | ")} |`
-      );
-
-      data.forEach((row) => {
-      const rowValues = headers.map((header, index) => {
-        const originalIndex = headerIndexMap[header];
-        return pad(row[originalIndex] || "", columnWidths[index]);
-      });
-      mdRows.push(`| ${rowValues.join(" | ")} |`);
-    });
-
-      return mdRows.join("\n");
-    },
     cellFormatter(cell, params, onRendered) {
       let cellVal = cell.getValue()
       if (!!cellVal && cellVal?.length > 1000) {
@@ -466,15 +358,24 @@ export default {
           },
           {
             label: '<i class="fas fa-copy"></i><span>Copy as JSON</span>',
-            action: () => this.copyTableData("json"),
+            action: () => {
+              const data = last(this.table.getRangesData());
+              this.copyTableData(data, "json", this.columns);
+            },
           },
           {
             label: '<i class="fas fa-copy"></i><span>Copy as CSV</span>',
-            action: () => this.copyTableData("csv"),
+            action: () => {
+              const data = last(this.table.getRangesData());
+              this.copyTableData(data, "csv", this.columns);
+            },
           },
           {
             label: '<i class="fas fa-copy"></i><span>Copy as Markdown</span>',
-            action: () => this.copyTableData("markdown"),
+            action: () => {
+              const data = last(this.table.getRangesData());
+              this.copyTableData(data, "markdown", this.columns);
+            },
           },
           {
             label: '<i class="fas fa-edit"></i><span>View Content</span>',
